Format the workspace and clear all clippy warnings
rustfmt across 19 files, then 29 clippy warnings to zero: 8 by
--fix, the rest by hand — type aliases for six complex types
(ComponentDef's boxed closures, DebugLine, DebugLine2D, DrawLineArgs),
&PathBuf to &Path in the CLI, struct-update syntax in Physics2DState,
and two identical branches merged in the Luau syntax highlighter.

No behaviour change. Tests pass.

+++ crates/balaur_core/src/script/det.rs
@@ -101,8 +101,8 @@ fn lua_random(eng: &Engine, args: Variadic<i64>) -> mlua::Result<Value> {
     let mut rng = rng.borrow_mut();
     match args.len() {
         0 => Ok(Value::Number(rng.0.next_f64())),
-        1 => Ok(Value::Integer(rng.0.next_range_i64(1, args[0]) as i64)),
-        _ => Ok(Value::Integer(rng.0.next_range_i64(args[0], args[1]) as i64)),
+        1 => Ok(Value::Integer(rng.0.next_range_i64(1, args[0]))),
+        _ => Ok(Value::Integer(rng.0.next_range_i64(args[0], args[1]))),
     }
 }
